refactor(server): merge engine module
