+++ engine/src/main/kotlin/Main.kt
@@ -1,6 +1,6 @@
 package dev.yidafu.blog
 
-import dev.yidafu.blog.dev.yidafu.blog.engine.*
+import dev.yidafu.blog.engine.*
 import org.koin.core.context.startKoin
 import org.koin.core.qualifier.StringQualifier
 import org.koin.ksp.generated.module
