+++ server/src/main/kotlin/dev/yidafu/blog/engine/ArticleManager.kt
@@ -1,7 +1,7 @@
-package dev.yidafu.blog.dev.yidafu.blog.engine
+package dev.yidafu.blog.engine
 
 import dev.yidafu.blog.common.dto.CommonArticleDTO
-import dev.yidafu.blog.dev.yidafu.blog.engine.TaskScope.Companion.NAME
+import dev.yidafu.blog.engine.TaskScope.Companion.NAME
 import org.koin.core.annotation.Scope
 import org.koin.core.annotation.Scoped
 import java.io.File
