+++ server/src/main/kotlin/dev/yidafu/blog/engine/CodeFenceGeneratingProvider.kt
@@ -1,4 +1,4 @@
-package dev.yidafu.blog.dev.yidafu.blog.engine
+package dev.yidafu.blog.engine
 
 import org.intellij.markdown.MarkdownTokenTypes
 import org.intellij.markdown.ast.ASTNode
