+++ server/src/main/kotlin/dev/yidafu/blog/engine/FileExt.kt
@@ -1,4 +1,4 @@
-package dev.yidafu.blog.dev.yidafu.blog.engine
+package dev.yidafu.blog.engine
 
 import java.nio.file.Path
 import java.time.LocalDateTime
