+++ server/src/main/kotlin/dev/yidafu/blog/engine/MarkdownExt.kt
@@ -1,4 +1,4 @@
-package dev.yidafu.blog.dev.yidafu.blog.engine
+package dev.yidafu.blog.engine
 
 import org.intellij.markdown.IElementType
 import org.intellij.markdown.ast.ASTNode
