+++ server/src/main/kotlin/dev/yidafu/blog/engine/processor/MarkdownProcessor.kt
@@ -1,10 +1,12 @@
-package dev.yidafu.blog.dev.yidafu.blog.engine.processor
+package dev.yidafu.blog.engine.processor
 
 import com.charleskorn.kaml.Yaml
 import dev.yidafu.blog.common.dto.CommonArticleDTO
 import dev.yidafu.blog.common.dto.FrontMatterDTO
 import dev.yidafu.blog.common.modal.ArticleSourceType
-import dev.yidafu.blog.dev.yidafu.blog.engine.*
+import dev.yidafu.blog.engine.*
+import dev.yidafu.blog.engine.CodeFenceGeneratingProvider
+import dev.yidafu.blog.engine.ImageGeneratingProvider
 import kotlinx.serialization.decodeFromString
 import org.intellij.markdown.IElementType
 import org.intellij.markdown.MarkdownElementTypes
